Life is present always.
Mind always tells  body will dissolved.   breath finish. It is true. 
Question comes in mind when we are living. seeing and realizes this soft, dreamy world and universe. What left and what we are doing in life. Making money and collects information in from of knowledge. 
Each life wants to live always with his life. We know it is possible for human. As we each time remember and feel  the presence of God/ Cosmic master/Guru. We worship and believe. One is whole master. 
Kriyayoga-Science is the path for only human. Time tells and we are human. 
Hindu/Muslim/sikh/Isai/Bodhi/Jaini etc.  are  religious wave.  
Acceptance is the key of wish. To practice for getting all in life.
Accepting is practice. Kriyayoga -Science is 100% easiest only one way path for human.